==> 1_External_Data_Importation.ipynb <==
import pandas as pd
import numpy as np
from matplotlib import pyplot as plt

### Get data from openstreetmap

osm_data = pd.read_csv('dataset/OSM.csv')

osm_data.head()

### Karena surface dan lit constant, kita drop saja

osm_data = osm_data.drop(['surface','lit'], axis = 1)

osm_data.head()

### Export aja karena yg paling penting maxspeednya

osm_data.to_csv(r'dataset/external_data.csv', index=False)
