==> 2_Data_Preparation.ipynb <==
import pandas as pd
import numpy as np
from matplotlib import pyplot as plt

df_train = pd.read_csv('dataset/train.csv')
df_test = pd.read_csv('dataset/test.csv')

df_train.head()

df_train.info()

## Timestamp formating

# Slice and format datatime 

df_train['waktu_setempat'] = df_train['waktu_setempat'].str.slice(0, 19)
df_train['waktu_setempat'] = pd.to_datetime(df_train['waktu_setempat'])

df_test['waktu_setempat'] = df_test['waktu_setempat'].str.slice(0, 19)
df_test['waktu_setempat'] = pd.to_datetime(df_test['waktu_setempat'])

df_train.head()

## External Data Merger

static_features = pd.read_csv('dataset/external_data.csv')
static_features.head()

df_train = df_train.merge(static_features, on='id_jalan', how='left')
df_test = df_test.merge(static_features, on='id_jalan', how='left')

## Ide: Membuat id baru untuk tiap id_jalan - id_titik_mulai - id_titik_akhir yang unique

list_id = ['id_jalan', 'id_titik_mulai', 'id_titik_akhir']

for i in list_id:
    df_train[i] = df_train[i].astype('str')
    df_test[i] = df_test[i].astype('str')

df_train['new_id'] = df_train['id_jalan'] + '_' + df_train['id_titik_mulai'] + '_' + df_train['id_titik_akhir']
df_train = df_train.drop(list_id, axis = 1)

df_test['new_id'] = df_test['id_jalan'] + '_' + df_test['id_titik_mulai'] + '_' + df_test['id_titik_akhir']
df_test = df_test.drop(list_id, axis = 1)

df_train.head()

len(df_train.new_id.unique())

Saat ini, terdapat 934 time series yang berbeda yang siap untuk diproses

### Save into csv

df_train.to_csv(r'dataset/train_prepared.csv', index=False)
df_test.to_csv(r'dataset/test_prepared.csv', index=False)
