==> 4_Data_Preprocessing.ipynb <==
import pandas as pd
import numpy as np
from matplotlib import pyplot as plt

df_train = pd.read_csv('dataset/train_prepared.csv')
df_train['waktu_setempat'] = pd.to_datetime(df_train['waktu_setempat'])
df_train = df_train.rename(columns={'waktu_setempat': 'timestamp', 'new_id': 'item_id'})

df_test = pd.read_csv('dataset/test_prepared.csv')
df_test['waktu_setempat'] = pd.to_datetime(df_test['waktu_setempat'])
df_test = df_test.rename(columns={'waktu_setempat': 'timestamp', 'new_id': 'item_id'})

## Null Handling and Skipping Timestamp Handling

### Null Handling

#Cek Null
print("df train\n", df_train.isna().sum(), "\n")
print("df test\n",df_test.isna().sum())

Dapat terlihat bahwa nilai null hanya terdapat pada feature lanes. Feature lanes merupakan feature yang menjelaskan jumlah lanes pada jalan tersebut. Berdasarkan intuisi dan domain knowledge, dapat diasumsikan bahwa apabila suatu lanes bernilai null, maka itu disebabkan karena hanya ada satu lanes pada jalan tersebut. Maka dari itu, nilai null pada feature tersebut akan diisi dengan nilai 1 

#fillna lanes
df_train['lanes'] = df_train['lanes'].fillna(1)
df_test['lanes'] = df_test['lanes'].fillna(1)

### Skipping Timestamp Handling

Pada kasus kali ini, digunakan metode filling data dari library autoML yaitu autogluon untuk handling skipping data (tanggal atau jam yang terlewat). Proses ini hanya perlu dilakukan pada train data

from autogluon.timeseries import TimeSeriesDataFrame, TimeSeriesPredictor

train_data = TimeSeriesDataFrame.from_data_frame(
    df_train,
    id_column="item_id",
    timestamp_column="timestamp"
)

train_data = train_data.to_regular_index(freq="H")
print(f"Data has frequency '{train_data.freq}'")

train_data = train_data.fill_missing_values(method="auto")

# Turn back into pandas dataframe

df_train = pd.DataFrame(train_data)
df_train = df_train.reset_index()

df_train.head()

## Categorical Feature Encoding

Pada kasus ini, terdapat satu feature yang bertipe data object yaitu highway. Maka dari itu akan dilakukan encoding berupa label encoding.

# Label Encode "highway"

ordinal_map = {'primary':2,
               'trunk':1,
               'secondary':0
}

df_train['highway'] = df_train.highway.map(ordinal_map)
df_test['highway'] = df_test.highway.map(ordinal_map)

### Save Data

df_train.to_csv(r'dataset/train_preprocessed.csv', index=False)
df_test.to_csv(r'dataset/test_preprocessed.csv', index=False)
