==> 7_Modelling_Autogluon_Ts.ipynb <==
import pandas as pd
import numpy as np
from matplotlib import pyplot as plt
from autogluon.timeseries import TimeSeriesDataFrame, TimeSeriesPredictor

df_train = pd.read_csv('dataset/train_feature_selected.csv')
df_train['timestamp'] = pd.to_datetime(df_train['timestamp'])

df_test = pd.read_csv('dataset/test_feature_selected.csv')
df_test['timestamp'] = pd.to_datetime(df_test['timestamp'])

df_test_ori = pd.read_csv('dataset/test.csv')

## Modeling

df_train.info()

train_data = TimeSeriesDataFrame.from_data_frame(
    df_train,
    id_column="item_id",
    timestamp_column="timestamp"
)

test_data = TimeSeriesDataFrame.from_data_frame(
    df_test,
    id_column="item_id",
    timestamp_column="timestamp"
)

predictor = TimeSeriesPredictor(
    prediction_length=200,
    path="models/autogluon-model",
    target="rerata_kecepatan",
    eval_metric="sMAPE",
    #random_seed = 42,
    #known_covariates_names=["mean_rerata_kecepatan_mingguan",'mean_rerata_kecepatan_harian','hour','2_per_3_maxspeed','highway'],
    #ignore_time_index=True
)

predictor.fit(
    train_data,
    presets="best_quality",
    time_limit=900
)

predictions = predictor.predict(train_data)
predictions.head()

### Predict Test Data

predictions = pd.DataFrame(predictions)

predictions.head()

### Create Submission file

df_test.head()

submission = pd.DataFrame()
submission['id'] = df_test_ori['id']
submission['rerata_kecepatan'] = 0

for i in submission['id']:
    time = df_test['timestamp'][i]
    series_id = df_test['item_id'][i]
    result = predictions.loc[(series_id, time)]['mean']
    submission.loc[i, 'rerata_kecepatan'] = result
    

submission.head()

submission.to_csv(r'submission/submission_new_2.csv', index=False)
